Restructure chunk_dispatch
This patch adjusts the code layout for chunk_dispatch to be similar
to the other custom nodes. All the files related to chunk_dispatch
are moved into dedicated nodes/chunk_dispatch directory.
