+++ src/copy.c
@@ -60,8 +60,8 @@
 #include "cross_module_fn.h"
 #include "dimension.h"
 #include "hypertable.h"
-#include "nodes/chunk_dispatch.h"
-#include "nodes/chunk_insert_state.h"
+#include "nodes/chunk_dispatch/chunk_dispatch.h"
+#include "nodes/chunk_dispatch/chunk_insert_state.h"
 #include "subspace_store.h"
 
 #if PG14_GE
