--- src/nodes/chunk_dispatch_plan.c
@@ -1,111 +1,0 @@
-/*
- * This file and its contents are licensed under the Apache License 2.0.
- * Please see the included NOTICE for copyright information and
- * LICENSE-APACHE for a copy of the license.
- */
-#include <postgres.h>
-#include <nodes/nodes.h>
-#include <nodes/extensible.h>
-#include <nodes/makefuncs.h>
-#include <nodes/nodeFuncs.h>
-#include <nodes/readfuncs.h>
-#include <parser/parsetree.h>
-#include <utils/rel.h>
-#include <catalog/pg_type.h>
-#include <rewrite/rewriteManip.h>
-
-#include "chunk_dispatch_plan.h"
-#include "chunk_dispatch_state.h"
-#include "hypertable.h"
-
-#include "compat/compat.h"
-
-/*
- * Create a ChunkDispatchState node from this plan. This is the full execution
- * state that replaces the plan node as the plan moves from planning to
- * execution.
- */
-static Node *
-create_chunk_dispatch_state(CustomScan *cscan)
-{
-	return (Node *) ts_chunk_dispatch_state_create(linitial_oid(cscan->custom_private),
-												   linitial(cscan->custom_plans));
-}
-
-static CustomScanMethods chunk_dispatch_plan_methods = {
-	.CustomName = "ChunkDispatch",
-	.CreateCustomScanState = create_chunk_dispatch_state,
-};
-
-/* Create a chunk dispatch plan node in the form of a CustomScan node. The
- * purpose of this plan node is to dispatch (route) tuples to the correct chunk
- * in a hypertable.
- *
- * Note that CustomScan nodes cannot be extended (by struct embedding) because
- * they might be copied, therefore we pass hypertable_relid in the
- * custom_private field.
- *
- * The chunk dispatch plan takes the original tuple-producing subplan, which
- * was part of a ModifyTable node, and imposes itself between the
- * ModifyTable plan and the subplan. During execution, the subplan will
- * produce the new tuples that the chunk dispatch node routes before passing
- * them up to the ModifyTable node.
- */
-static Plan *
-chunk_dispatch_plan_create(PlannerInfo *root, RelOptInfo *relopt, CustomPath *best_path,
-						   List *tlist, List *clauses, List *custom_plans)
-{
-	ChunkDispatchPath *cdpath = (ChunkDispatchPath *) best_path;
-	CustomScan *cscan = makeNode(CustomScan);
-	ListCell *lc;
-
-	foreach (lc, custom_plans)
-	{
-		Plan *subplan = lfirst(lc);
-
-		cscan->scan.plan.startup_cost += subplan->startup_cost;
-		cscan->scan.plan.total_cost += subplan->total_cost;
-		cscan->scan.plan.plan_rows += subplan->plan_rows;
-		cscan->scan.plan.plan_width += subplan->plan_width;
-	}
-
-	cscan->custom_private = list_make1_oid(cdpath->hypertable_relid);
-	cscan->methods = &chunk_dispatch_plan_methods;
-	cscan->custom_plans = custom_plans;
-	cscan->scan.scanrelid = 0; /* Indicate this is not a real relation we are
-								* scanning */
-	/* The "input" and "output" target lists should be the same */
-	cscan->custom_scan_tlist = tlist;
-	cscan->scan.plan.targetlist = tlist;
-
-	return &cscan->scan.plan;
-}
-
-static CustomPathMethods chunk_dispatch_path_methods = {
-	.CustomName = "ChunkDispatchPath",
-	.PlanCustomPath = chunk_dispatch_plan_create,
-};
-
-Path *
-ts_chunk_dispatch_path_create(PlannerInfo *root, ModifyTablePath *mtpath, Index hypertable_rti,
-							  int subpath_index)
-{
-	ChunkDispatchPath *path = (ChunkDispatchPath *) palloc0(sizeof(ChunkDispatchPath));
-#if PG14_LT
-	Path *subpath = list_nth(mtpath->subpaths, subpath_index);
-#else
-	Path *subpath = mtpath->subpath;
-#endif
-	RangeTblEntry *rte = planner_rt_fetch(hypertable_rti, root);
-
-	memcpy(&path->cpath.path, subpath, sizeof(Path));
-	path->cpath.path.type = T_CustomPath;
-	path->cpath.path.pathtype = T_CustomScan;
-	path->cpath.methods = &chunk_dispatch_path_methods;
-	path->cpath.custom_paths = list_make1(subpath);
-	path->mtpath = mtpath;
-	path->hypertable_rti = hypertable_rti;
-	path->hypertable_relid = rte->relid;
-
-	return &path->cpath.path;
-}
