+++ tsl/src/nodes/data_node_dispatch.c
@@ -22,9 +22,8 @@
 
 #include "ts_catalog/chunk_data_node.h"
 
-#include <nodes/chunk_dispatch_plan.h>
-#include <nodes/chunk_dispatch_state.h>
-#include <nodes/chunk_insert_state.h>
+#include <nodes/chunk_dispatch/chunk_dispatch.h>
+#include <nodes/chunk_dispatch/chunk_insert_state.h>
 #include <hypertable_cache.h>
 #include <compat/compat.h>
 #include <guc.h>
