+++ tsl/src/remote/dist_copy.c
@@ -25,8 +25,8 @@
 #include "guc.h"
 #include "hypercube.h"
 #include "hypertable.h"
-#include "nodes/chunk_dispatch.h"
-#include "nodes/chunk_insert_state.h"
+#include "nodes/chunk_dispatch/chunk_dispatch.h"
+#include "nodes/chunk_dispatch/chunk_insert_state.h"
 #include "partitioning.h"
 #include "remote/connection_cache.h"
 #include "remote/dist_txn.h"
